(#79) 🔄 refactor: 챗 컴포넌트는 하나의 디렉토리 안으로 정리
